+++ package.json
@@ -11,7 +11,7 @@
     "lodash": "4.2.1",
     "multer": "1.1.0",
     "sqlite3": "3.1.1",
-    "substance": "substance/substance#cbe49001e833190bc0688495f28ec083f8cb18f6",
+    "substance": "substance/substance#2f7539be75e798d8b6edaa9289ac581eb920a943",
     "ws": "^1.0.1"
   },
   "devDependencies": {
