+++ server.js
@@ -8,14 +8,11 @@ var bodyParser = require('body-parser');
 var http = require('http');
 var WebSocketServer = require('ws').Server;
 
-var figures = require('./hub/figures');
-var snapshots = require('./hub/snapshots');
-
 var knexConfig = require('./knexfile');
 var port = process.env.PORT || 5000;
 var host = process.env.HOST || 'localhost';
 var wsUrl = process.env.WS_URL || 'ws://'+host+':'+port;
-var backend = new Backend({config: knexConfig});
+var backend = new Backend({config: knexConfig}, require('./note/Note.js'));
 
 // If seed option provided we should remove db, run migration and seed script
 if(process.argv[2] == 'seed') {
@@ -59,11 +56,6 @@ var wss = new WebSocketServer({ server: httpServer });
 var hub = new CollabHub(wss, backend);
 // Adds http routes that CollabHub implements
 hub.addRoutes(app);
-
-snapshots.connect(backend);
-snapshots.addRoutes(app);
-
-figures.addRoutes(app);
 
 // Delegate http requests to express app
 httpServer.on('request', app);
